+++ FightingFantasy.Mvc/ApiClients/ApiClient.cs
@@ -1304,10 +1304,10 @@ namespace FightingFantasy.Mvc.ApiClients
         public System.Collections.Generic.ICollection<PlaythroughStatModel> Stats { get; set; }
     
         [Newtonsoft.Json.JsonProperty("xPos", Required = Newtonsoft.Json.Required.DisallowNull, NullValueHandling = Newtonsoft.Json.NullValueHandling.Ignore)]
-        public int XPos { get; set; }
+        public double XPos { get; set; }
     
         [Newtonsoft.Json.JsonProperty("yPos", Required = Newtonsoft.Json.Required.DisallowNull, NullValueHandling = Newtonsoft.Json.NullValueHandling.Ignore)]
-        public int YPos { get; set; }
+        public double YPos { get; set; }
     
     
     }
